py3k: print_function, absolute_import, range
Running python-modernize to automatically fix these problems.
To avoid dependency on the 'six' package, range and zip have been
left as is. On Python 3 these return iterators, while they return
lists in Python 2.
We accept that there will be more memory used in Python 2, and have
only made sure that if a list is expected, we wrap it explicitly.